Fix up in-flight errors for 1.2.
With the push for 1.2, we had a bunch of stuff happening in parallel,
that made some issues when they all landed. Notably, we changed the type
being returned by ProviderConfig, which made every use of meta that
wasn't updated panic. All the uses that were merged at the time were
updated, but the ones in flight when the PR was opened were not, causing
them to panic. This PR fixes that.

The vendor for `jobspec` from Nomad was also updated, as it didn't have
a field for namespace at the time it was vendored, which means that jobs
could not be scheduled in namespaces. Now they can.