Resolved. Notice when a Nomad job has changed.
Resolved #1. Terraform fails to notice when a Nomad job has changed
To have this, you need to maximize the use of Job Meta. (eg: Docker
Image)
This patch only notice when number of running instances of group in job
has changed.
Tested with Nomad 0.11.3
Tested with `Service` job.
Batch is assumed to be save to run multiple times so there would be no
problem.