provider: support TLS certs configured by strings.
This change adds a number of new provider config params that allow
users to configure TLS certificates via strings, alongside the
pre-existing file location method.

The provider schema ConflictsWith param is used to notify users
when a configuration attempts to use two keys which affect the
same TLS config param. In order to successfully use this, the
existing DefaultFunc default values have been moved from an empty
string to nil. This is because Terraform treats an empty string
var as being populated, and therefore raises a conflict error
whenever a new param is set. I tested this change locally to ensure
it didn't cause any compatibility issues, and didn't find any. It
is reassures that the Consul provider went through the same change.