provider: don't load namespace from env var
The Nomad provider doesn't have a global Namespace configuration as each
resource may target a different namespace, but the Nomad API client
default configuration would load a namespace value from the
NOMAD_NAMESPACE automatically.

This caused issues in environments where Terraform runs within a Nomad
allocation, where the value for NOMAD_NAMESPACE is the namespace of the
allocation.

Since the namespace value was never supposed to be set at the provider
level, this change makes sure the Nomad client namespace is always
empty.