provider: fix namespaced importer
Importing namespaced resources require a namespace to be passed. Without
this requirement jobs such as `a@b` become ambiguous as it's not
possible to determine if the user intention was to import the job named
`a` from namespace `b` or job `a@b` on namespace `default`. By requiring
a namespace to be set this ambiguity is solved.

Additionally, Terraform automatically refreshes the state after an
import, so there is no need to call the `Read` method manually on
import. Doing so can also have unintended consequences, as the provider
deletes resources from state in case of a `404`, resulting in a
Terraform error when trying to import resources that don't exist.