job: import jobs from non-default namespace (#408)
* import: add helper to import namespaced resources

`nomad_job`, `nomad_csi_volume`, and `nomad_csi_volume_registration` do
not have the namespace as part of their state ID. This makes it
impossible to import them if they are registered in a non-default
namespace since only the ID is provided to the import function.

This commit adds a new helper that imports using an ID with the format
`<id>@<namespace>`. Since `@` is not a valid character for namespaces,
the last `@` character (if found) represents the separator.

* changelog: add entry for #408

* chore: add missing copyright header

* provider: fix namespaced importer

Importing namespaced resources require a namespace to be passed. Without
this requirement jobs such as `a@b` become ambiguous as it's not
possible to determine if the user intention was to import the job named
`a` from namespace `b` or job `a@b` on namespace `default`. By requiring
a namespace to be set this ambiguity is solved.

Additionally, Terraform automatically refreshes the state after an
import, so there is no need to call the `Read` method manually on
import. Doing so can also have unintended consequences, as the provider
deletes resources from state in case of a `404`, resulting in a
Terraform error when trying to import resources that don't exist.

* apply code review