testing: fix test flake in data.nomad_allocations test (#502)
The acceptance test for `data.nomad_allocations` creates a job and then reads
the allocations in the same configuration. This results in frequent test flakes
because registering a job via `resource.nomad_job` will typically return before
the allocations have actually been created, making the `depends_on` useless.

This frankly isn't a good workflow to encourage, as allocations are
ephemeral. Fix the test by creating the job first, waiting for the allocations
to be live, and then adding the `data.nomad_allocations` object to the config.