auth method: fix handling of redacted OIDC client secret field
In Nomad 1.9.7 we started redacting the OIDC client secret field from the Auth
Method API. Once we update the Nomad test target, this results in incorrect plan
diffs. But we don't want to simply suppress the diff of this field, because we
need to record the original unredacted value in state. Update the resource to
record the original secret to state and then use the state to override the
value we get from the API when performing the diff.

Note that it's impossible with or without this patch to import an Auth Method
without manual intervention, because we can't create the OIDC client secret from
nothing.

Ref: https://github.com/hashicorp/nomad/pull/25328